Reordenació dels fitxers en els directoris originals
