+++ README.md
@@ -5,7 +5,7 @@ Objectiu
 --------
 
 Aprenentatge de procediments relacionats amb el desenvolupament de programari:
-- Ús dels IDEs, refactorització i generació automàtica de codi. 
+- Ús dels IDEs, refactorització i generació automàtica de codi.
 - Proves automatitzades.
 - Git.
 - UML.
@@ -14,3 +14,8 @@ Tecnologies
 -----------
 
 - Java
+
+Diagrames
+---------
+
+![Diagrama de classes i seqüència de Valoració dels comentaris](./docs/Diagrames-ValoracioComentaris.png)
